Extract FlagValue and FlagValueType files
